+++ server/tests-hspec/Harness/Test/Context.hs
@@ -1,6 +1,9 @@
-{-# LANGUAGE DeriveAnyClass #-}
-
--- | Helper functions for easily testing features.
+-- | Helper functions for testing in specific contexts.
+--
+--   A 'Context' represents the prerequisites for running a test,
+--   such as the required backend, the setup process of tables and permissions,
+--   the creation of local state and the teardown of created context after the test
+--   is done.
 module Harness.Test.Context
   ( run,
     runWithLocalState,
@@ -14,9 +17,9 @@ module Harness.Test.Context
 where
 
 import Control.Applicative ((<|>))
-import Control.Exception.Safe (Exception, SomeException, catch, mask, throwIO)
 import Data.Foldable (for_)
 import Data.Maybe (fromMaybe)
+import Harness.Exceptions (catchRethrow, mask)
 import Harness.State (State)
 import Test.Hspec (ActionWith, HasCallStack, SpecWith, aroundAllWith, describe)
 import Test.Hspec.Core.Spec (Item (..), mapSpecItem)
@@ -79,7 +82,7 @@ mapItemAction mapActionWith item@Item {itemExample} =
 
 -- | Runs the given tests, for each provided 'Context'@ a@.
 --
--- Each 'Context' provides distinct setup and teardown functions; 'feature'
+-- Each 'Context' provides distinct setup and teardown functions; 'runWithLocalState'
 -- guarantees that the associated 'teardown' function is always called after a
 -- setup, even if the tests fail.
 --
@@ -101,9 +104,8 @@ runWithLocalState contexts tests =
     -- and at teardown, which is why we use a custom re-implementation of
     -- @bracket@.
     contextBracket ::
-      forall b.
       Context a ->
-      ((State, a) -> IO b) ->
+      ((State, a) -> IO ()) ->
       State ->
       IO ()
     contextBracket Context {mkLocalState, setup, teardown} actionWith globalState =
@@ -111,31 +113,15 @@ runWithLocalState contexts tests =
         localState <- mkLocalState globalState
         let state = (globalState, localState)
 
-        catch
-          -- Setup for a test
+        catchRethrow
           (setup state)
-          ( \setupEx ->
-              catch
-                -- On setup error, attempt to run `teardown` and then throw the setup error
-                (teardown state *> throwIO setupEx)
-                -- On teardown error as well, throw both exceptions
-                (throwIO . Exceptions setupEx)
-          )
+          (teardown state)
 
         -- Run tests.
         _ <-
-          catch
+          catchRethrow
             (restore $ actionWith state)
-            ( \restoreEx -> do
-                -- On test error, attempt to run `teardown`...
-                teardown state
-                  `catch`
-                  -- ...if it fails as well, bundle both exceptions together and
-                  -- rethrow them...
-                  (throwIO . Exceptions restoreEx)
-                -- ...otherwise rethrow the original exception.
-                throwIO restoreEx
-            )
+            (teardown state)
         -- If no exception occurred, run the normal teardown function.
         teardown state
 
@@ -171,7 +157,7 @@ runWithLocalState contexts tests =
 data Context a = Context
   { -- | A name describing the given context.
     --
-    -- e.g. @Postgre@ or @MySQL@
+    -- e.g. @Postgres@ or @MySQL@
     name :: ContextName,
     -- | Setup actions associated with creating a local state for this 'Context'; for example:
     --  * starting remote servers
@@ -251,19 +237,3 @@ defaultOptions =
   Options
     { stringifyNumbers = False
     }
-
---------------------------------------------------------------------------------
--- Local helpers
-
--- | Two exceptions, bundled as one.
-data Exceptions
-  = Exceptions SomeException SomeException
-  deriving anyclass (Exception)
-
-instance Show Exceptions where
-  show (Exceptions e1 e2) =
-    unlines
-      [ "1. " <> show e1,
-        "",
-        "2. " <> show e2
-      ]