+++ CHANGELOG.md
@@ -5,10 +5,19 @@
 ### Bug fixes and improvements
 (Add entries below in the order of server, console, cli, docs, others)
 
+- server: add custom function for case insensitive lookup in session variable in request transformation
 - server: Webhook Transforms can now delete request/response bodies explicitly.
 - server: Fix truncation of session variables with variable length column types in MSSQL (#8158)
+- server: improve performance of `replace_metadata` for large schemas
 - server: improve baseline memory consumption for typical workloads
 - server: fix parsing timestamp values in BigQuery backends (fix #8076)
+- console: include cron trigger with include in metadata as false on cron trigger manage page
+- cli: fix metadata version being set to 3 when doing `hasura init --version 2` (#8148)
+
+## v2.2.1
+
+- server: postgres: return a single entry per row (selected randomly) when an object relationship has multiple matches (fix #7936)
+- console: add support for remote database relationships
 
 ## v2.3.0-beta.2
 
@@ -44,7 +53,7 @@ The optimization can be enabled using the
   update mutations.
 - server: allow inserting more than 1 row simultaneously into table with generated columns (fix #4633)
   that have generated columns in Postgres.
-- server: postgres: return a single entry per row (selected randomly) when an object relationship is misconfigured one-to-many
+- server: postgres: return a single entry per row (selected randomly) when an object relationship has multiple matches (fix #7936)
 - server: Updates Kriti to v0.3.0
 - server: add operation name in the request sent to remote schemas
 - server: add support for scalar response types for actions (fix #7805)